Fix potential segfaults in test cases
In many places checking for nullptr is done using EXPECT test macros
that doesn't return from the function when condition is not met
(i.e. they are non-fatal). Lack of further error handling results
in a subsequent segfault. To avoid that in most places it's sufficient
to change macro type to ASSERT.

Signed-off-by: Robert Drab <robert.drab@huawei.com>